update: 完成基本功能

+++ src/bot/handlers/command_admin.rs
@@ -1,4 +1,4 @@
-use anyhow::{anyhow, Result};
+use anyhow::{Context, Result};
 use teloxide::dispatching::DpHandlerDescription;
 use teloxide::dptree::case;
 use teloxide::prelude::*;
@@ -10,8 +10,8 @@ use crate::bot::filter::filter_admin_msg;
 use crate::bot::Bot;
 use crate::database::{GalleryEntity, MessageEntity};
 use crate::ehentai::EhGalleryUrl;
-use crate::manager::uploader::ExloliUploader;
 use crate::reply_to;
+use crate::uploader::ExloliUploader;
 
 pub fn admin_command_handler() -> Handler<'static, DependencyMap, Result<()>, DpHandlerDescription>
 {
@@ -19,7 +19,7 @@ pub fn admin_command_handler() -> Handler<'static, DependencyMap, Result<()>, Dp
         .chain(filter_admin_msg())
         .branch(case![AdminCommand::Upload(gallery)].endpoint(cmd_upload))
         .branch(case![AdminCommand::Delete].endpoint(cmd_delete))
-        .branch(case![AdminCommand::Erase].endpoint(cmd_erase))
+        .branch(case![AdminCommand::Erase].endpoint(cmd_delete))
 }
 
 async fn cmd_upload(
@@ -28,27 +28,29 @@ async fn cmd_upload(
     uploader: ExloliUploader,
     gallery: EhGalleryUrl,
 ) -> Result<()> {
+    info!("{:?}: /upload {}", msg.from(), gallery);
     let reply = reply_to!(bot, msg, "上传中……").await?;
-    uploader.check_and_upload(&gallery).await?;
+    uploader.try_upload(&gallery).await?;
     bot.edit_message_text(msg.chat.id, reply.id, "上传完成").await?;
     Ok(())
 }
 
-async fn cmd_delete(bot: Bot, msg: Message) -> Result<()> {
-    let reply_to = msg.reply_to_message().ok_or(anyhow!("No reply message"))?;
-    let channel = reply_to.forward_from_chat().ok_or(anyhow!("No forward from chat"))?;
-    let msg_entity = MessageEntity::get(reply_to.id.0).await?.unwrap();
-    bot.delete_message(channel.id, MessageId(msg_entity.id)).await?;
-    GalleryEntity::update_deleted(msg_entity.gallery_id, true).await?;
-    Ok(())
-}
+async fn cmd_delete(bot: Bot, msg: Message, command: AdminCommand) -> Result<()> {
+    let reply_to = msg.reply_to_message().context("没有回复消息")?;
 
-async fn cmd_erase(bot: Bot, msg: Message) -> Result<()> {
-    let reply_to = msg.reply_to_message().ok_or(anyhow!("No reply message"))?;
-    let channel = reply_to.forward_from_chat().ok_or(anyhow!("No forward from chat"))?;
-    let msg_entity = MessageEntity::get(reply_to.id.0).await?.unwrap();
+    let channel = reply_to.forward_from_chat().context("该消息没有回复画廊")?;
+    let channel_msg = reply_to.forward_from_message_id().context("获取转发来源失败")?;
+
+    let msg_entity = MessageEntity::get(channel_msg).await?.unwrap();
+
     bot.delete_message(reply_to.chat.id, reply_to.id).await?;
     bot.delete_message(channel.id, MessageId(msg_entity.id)).await?;
-    GalleryEntity::delete(msg_entity.gallery_id).await?;
+
+    if matches!(command, AdminCommand::Delete) {
+        GalleryEntity::update_deleted(msg_entity.gallery_id, true).await?;
+    } else {
+        GalleryEntity::delete(msg_entity.gallery_id).await?;
+    }
+
     Ok(())
 }
